+++ packages/stream-tile-handler/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.3.0-rc.0](/compare/@ceramicnetwork/stream-tile-handler@2.2.8...@ceramicnetwork/stream-tile-handler@2.3.0-rc.0) (2022-09-28)
+
+
+### Features
+
+* update dids, add/register cacao verifiers (#2452) d93fedb, closes #2452
+
+
+
+
+
 ## [2.2.8](/compare/@ceramicnetwork/stream-tile-handler@2.2.8-rc.3...@ceramicnetwork/stream-tile-handler@2.2.8) (2022-09-21)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-tile-handler